requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
dj-rest-auth==7.0.1
django-allauth==65.13.0
python-dotenv==1.0.0
dj-database-url==2.1.0
requests==2.32.5
django-cors-headers==4.3.1


